requirements.txt
Django==1.9.4
django-webpack-loader
djangorestframework
python-social-auth
django-oauth-toolkit
django-rest-framework-social-oauth2
django-settings-export
